remove statsdb
search for state when not a standard one
