+++ tests/cli/data/test_pocount_po_fuzzy/stdout.txt
@@ -5,6 +5,6 @@ Fuzzy:            1 ( 50%)          2 ( 50%)             n/a
 Untranslated:     0 (  0%)          0 (  0%)             n/a
 Total:            2                 4                      3
 
-Needs-Work:       1 ( 50%)          2 ( 50%)               0
 Unreviewed:       1 ( 50%)          2 ( 50%)               3
+Needs-Work:       1 ( 50%)          2 ( 50%)               0
