+++ translate/convert/convert.py
@@ -538,12 +538,12 @@ def should_output_store(store, threshold):
     if not threshold:
         return True
 
-    from translate.storage import statsdb
+    from translate.tools import pocount
 
     units = list(filter(lambda unit: unit.istranslatable(), store.units))
     translated = list(filter(lambda unit: unit.istranslated(), units))
     wordcounts = dict(
-        map(lambda unit: (unit.getid(), statsdb.wordsinunit(unit)), units)
+        map(lambda unit: (unit.getid(), pocount.wordsinunit(unit)), units)
     )
 
     def sourcewords(elementlist):
